Refactor health check endpoints and improve server shutdown handling: restore health check route, remove redundant status endpoints, and implement graceful shutdown logic
