+++ src/index.ts
@@ -1,5 +1,4 @@
 import "reflect-metadata"; // We need this in order to use @Decorators
-
 import express from "express";
 import Logger from "@/logger/logger";
 import config from "@/config/environment";
@@ -14,19 +13,30 @@ async function startServer() {
     // eslint-disable-next-line @typescript-eslint/no-require-imports
     await require("./loaders").default({ expressApp: app });
 
-    app
-      .listen(config.port, () => {
-        Logger.info(`
+    const server = app.listen(config.port, () => {
+      Logger.info(`
         ################################################
         🛡️  Server listening on port: ${config.port} 🛡️
         ################################################
       `);
-      })
-      .on("error", (err) => {
-        console.error("Server start error:", err);
-        Logger.error(err);
-        process.exit(1);
+    });
+
+    // Handle graceful shutdown
+    const shutdown = async () => {
+      console.log("\nReceived shutdown signal, shutting down gracefully...");
+      server.close((err) => {
+        if (err) {
+          console.error("Error during server shutdown:", err);
+          process.exit(1);
+        }
+        console.log("Server closed.");
+        // Close other resources like database connections here
+        process.exit(0);
       });
+    };
+
+    process.on("SIGTERM", shutdown);
+    process.on("SIGINT", shutdown);
   } catch (error) {
     console.error("Catastrophic server initialization error:", error);
     process.exit(1);
